backend/requirements.txt
Django==5.1.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers==4.3.1
django-redis==5.4.0
django-admin-sortable2==2.2.2
django-ckeditor-5==0.2.19
django-jazzmin==3.0.1
